Reorganização completa de componentes compartilhados e módulo "Home"
