+++ src/modules/boletas/screens/compra/compra_GainReducao/FormInternoCompraGainReducao.js
@@ -11,8 +11,8 @@ import {
   mudarAssinaturaAction,
   mudarCheckSalvarAssinaturaAction,
 } from "modules/boletas/duck/actions/formInputActions";
-import RowFormValidade from "shared/componentes/RowFormValidade";
-import RowFormAssinatura from "shared/componentes/RowFormAssinatura";
+import RowFormValidade from "modules/boletas/components/RowFormValidade";
+import RowFormAssinatura from "modules/boletas/components/RowFormAssinatura";
 import { COMPRA_GAINREDUCAO_NAMESPACE } from "constants/ActionTypes";
 import TabelaGainReducao from "./TabelaGainReducao";
 import { compraGainReducaoAction } from "modules/boletas/duck/actions/AppBoletasActions";
