+++ src/modules/home/screens/rightSideMenu/Alert/Alert.tsx
